Passing tests

+++ src/support/test/bit_cast.cpp
@@ -139,10 +139,10 @@ define signext i8 @value() {
   SECTION("more complicated for ints")
   {
     auto str = R"(
-define i32 @work(i32, i32, i32) {
-  %4 = mul nsw i32 %2, %1
-  %5 = add nsw i32 %4, %0
-  ret i32 %5
+define i64 @work(i64, i64, i64) {
+  %4 = mul nsw i64 %2, %1
+  %5 = add nsw i64 %4, %0
+  ret i64 %5
 })";
 
     PARSE_TEST_MODULE(mod, str);
@@ -152,7 +152,7 @@ define i32 @work(i32, i32, i32) {
     cb.add(2ll, 3ll, 4ll);
 
     auto ret = wrap.call(cb);
-    REQUIRE(bit_cast<int>(ret) == 14);
+    REQUIRE(bit_cast<int64_t>(ret) == 14);
   }
 
   SECTION("more complicated for floats")
